votes limit 1 per user per event and charities limit to 3

+++ resources/views/events/current.blade.php
@@ -5,21 +5,27 @@
 	<h1>{{ $currentEvent->title }}</h1>
 
 	@foreach ($currentEvent->charities as $charity)
+
 	<div class='charity'>
-		<p></p>
-		<p><a href="/charities/{{ $charity->id }}/delete" class="btn btn-danger pull-right"> Delete</a><a href="/charities/{{ $charity->id }}/edit" class="btn btn-warning pull-right" >Edit</a></p>
+		@if (Auth::user()->admin == 'admin')
+			<p><a href="/charities/{{ $charity->id }}/delete" class="btn btn-danger pull-right"> Delete</a><a href="/charities/{{ $charity->id }}/edit" class="btn btn-warning pull-right" >Edit</a></p>
+		@endif
 		<h4> {{ $charity->title }} </h4>
 		<p> {{ $charity->body }} </p>
-		<p>Total Votes: {{ $charity->votes->count() }}</p>
+		@if($currentEvent->votes()->count() != 0)
+			<p>Total Votes: {{ $charity->votes->count() }}</p>
+		@endif
 		<!-- button for voting  -->
-		<form method="POST" action="/charities/{{ $charity->id }}/votes">
-			<div class="form-group">
-				{{ csrf_field() }}
-				<a class="btn btn-primary" type="button" href="/charities/{{ $charity->id }}/voting">Vote</a>
-			</div>
+		@if($currentEvent->votes()->where('votes.user_id', $user)->count() == 0)
+			<form method="POST" action="/charities/{{ $charity->id }}/votes">
+				<div class="form-group">
+					{{ csrf_field() }}
+					<a class="btn btn-primary" type="button" href="/charities/{{ $charity->id }}/voting">Vote</a>
+				</div>
+		@endif
+
 		</form>
-	</div>	
-		
+	</div>
 		
 	@endforeach
